Add styles directory and move all styles to the styles directory.
